server/pyproject.toml
[project]
name = "invoice-reader"
version = "0.1.0"
description = "Manage your invoices."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic~=1.14.0",
    "bcrypt==4.0.1",
    "boto3~=1.35.56",
    "fastapi[standard]~=0.115.4",
    "loguru>=0.7.3",
    "mypy~=1.14.1",
    "passlib[bcrypt]~=1.7.4",
    "pdf2image>=1.17.0",
    "pillow>=11.2.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psycopg2-binary~=2.9.10",
    "pydantic~=2.9.2",
    "pydantic-settings>=2.10.1",
    "pyjwt~=2.10.0",
    "python-multipart~=0.0.17",
    "sqlmodel~=0.0.22",
    "together>=1.5.17",
]

[dependency-groups]
dev = [
    "ipykernel~=6.29.5",
    "pytest-mock~=3.14.0",
    "pytest~=8.3.3",
    "python-dotenv~=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100

[tool.ruff.format]
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = 20

[tool.ruff.lint]
extend-select = ["E501"]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
fixable = ["E", "F", "I"]

[tool.pylint.master]
disable = ["all"]
enable = [
    "missing-kwo-arg", 
    "unexpected-keyword-arg", 
    "no-value-for-parameter",
]

[tool.mypy]
strict = true
